+++ src/various_calculations/arithmetic_avrg.py
